Updated player data and added conversion

+++ Assets/Scripts/Networking/Game/Resource.cs
@@ -2,16 +2,19 @@ using DarkRift;
 
 namespace Networking.Game
 {
+    /// <summary>
+    /// 
+    /// </summary>
     public class Resource : IDarkRiftSerializable
     {
-        public ushort Id { get; set; }
+        public byte Id { get; set; }
         public string Name { get; set; }
         public ushort ConversionRate { get; set; }
-        public ushort Count { get; set; }
+        public uint Count { get; set; }
 
         public Resource() {}
         
-        public Resource(ushort id, string name, ushort conversionRate, ushort count)
+        public Resource(byte id, string name, ushort conversionRate, uint count)
         {
             Id = id;
             Name = name;
@@ -25,10 +28,10 @@ namespace Networking.Game
         /// <param name="e">Deserialize event</param>
         public void Deserialize(DeserializeEvent e)
         {
-            Id = e.Reader.ReadUInt16();
+            Id = e.Reader.ReadByte();
             Name = e.Reader.ReadString();
             ConversionRate = e.Reader.ReadUInt16();
-            Count = e.Reader.ReadUInt16();
+            Count = e.Reader.ReadUInt32();
         }
 
         /// <summary>
